Time: 55 ms (82.01%), Space: 18.8 MB (71.59%) - LeetHub

+++ 875-koko-eating-bananas/875-koko-eating-bananas.cpp
@@ -5,10 +5,10 @@ public:
     int minEatingSpeed(vector<int> &v, int h) {
      
         long long l = 1, r = *max_element(v.begin(),v.end());
-        int ans =r;
+        long long ans =r;
         while (l <= r) {
             long long sum = 0;
-            int m = (l+r)/2;
+            long long m = (l+r)/2;
             for(int i=0;i<v.size();++i){
                 sum+=(ceil(1.0*v[i]/m));
             }
